requirements.txt
apache-airflow==2.0.1
boto3==1.15
apache-airflow-providers-amazon==1.2.0
sqlalchemy==1.3.24
pytest==6.2.3
moto==2.0.8
freezegun